feat: use a single layout with multiple page types
having multiple layouts caused "flickering" when navigating
between different sections of the website because the browser
would reload the Layout component containing the navbar etc.

Now we're using a more simplified version of "pageTypes" that are based
on a custom Page component, resulting in more pleasing navigation
between pages.

This also has the benefit of making it easier to spot linking defects
(placing a standard <a href> link instead of using a router-link)
that trigger full page refreshes with the corresponding visual artifacts