fix: bundle size expectations
sadly the block index is now part of the main bundle since its part of
a vue plugin - apparently vite was only able to move it out of the
main bundle as long as the MaturityModel was only loaded as a component.

not sure this is a problem though - most pages will need the page index
(everything except the home page) and it doesn't blow up the main
bundle size as much either.