feat: select tool in cfmm via query param
by visiting a path like /maturity-model/?selectedTool=meshStack the user
will be shown the maturity model with the given tool selected. This is
useful to advertise a tool in a blog post: we can include a link in
the blog post which shows the user which parts of the cfmm are covered
by this particular tool.